feat: Tidying up nav initialization
