+++ src/Uno.Extensions.Navigation.UI/ServiceCollectionExtensions.cs
@@ -87,53 +87,6 @@ public static class ServiceCollectionExtensions
 #pragma warning restore CS8603 // Possible null reference return.
 	}
 
-	public static void AddInstance<T>(this IServiceProvider provider, Func<T> instanceCreator)
-	{
-		provider.AddInstance(typeof(T), instanceCreator);
-	}
-
-	public static void AddInstance<T>(this IServiceProvider provider, Type serviceType, Func<T> instanceCreator)
-	{
-		provider.GetRequiredService<IInstanceRepository>().Instances[serviceType] = instanceCreator;
-	}
-
-	public static T AddInstance<T>(this IServiceProvider provider, T instance)
-	{
-#pragma warning disable CS8604 // Possible null reference argument.
-		provider.AddInstance(typeof(T), instance);
-#pragma warning restore CS8604 // Possible null reference argument.
-		return instance;
-	}
-
-	public static object AddInstance(this IServiceProvider provider, Type serviceType, object instance)
-	{
-		provider.GetRequiredService<IInstanceRepository>().Instances[serviceType] = instance;
-		return instance;
-	}
-
-	public static T? GetInstance<T>(this IServiceProvider provider)
-	{
-		var value = provider.GetInstance(typeof(T));
-		if (value is Func<T> valueCreator)
-		{
-			var instance = valueCreator();
-			provider.AddInstance(instance);
-			return instance;
-		}
-
-		if (value is T typedValue)
-		{
-			return typedValue;
-		}
-
-		return default;
-	}
-
-	public static object? GetInstance(this IServiceProvider provider, Type type)
-	{
-		return provider.GetRequiredService<IInstanceRepository>().Instances.TryGetValue(type, out var value) ? value : null;
-	}
-
 	public static IServiceCollection AddRegion<TControl, TRegion>(this IServiceCollection services, string? name = null)
 		where TRegion : class, INavigator
 	{
@@ -150,4 +103,5 @@ public static class ServiceCollectionExtensions
 					.AddTransient<TData>(services => services.GetRequiredService<NavigationDataProvider>().GetData<TData>());
 #pragma warning restore CS8603 // Possible null reference return.
 	}
+
 }
